fix: use view! indentation settings when formatting syn expression (#141)
BREAKING CHANGE: When the indentation style is set to "Tabs", the formatter will print `pending_indentation / 4` hard tabs and the remainder (`pending_indentation % 4`) in spaces.  

The `tab_spaces` configuration property specifies the indentation size _in spaces_. This means that if you want an indentation of 1 hard tab (`\t`), you need to set `indentation_style = "Tabs"` and `tab_spaces = 4` (which is already the default).